[API] Convert Country and Province resource configs to APIP 3
