+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_platform/properties/Country.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<properties xmlns="https://api-platform.com/schema/metadata/properties-3.0"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="https://api-platform.com/schema/metadata/properties-3.0 https://api-platform.com/schema/metadata/properties-3.0.xsd"
+>
+    <property resource="Sylius\Component\Addressing\Model\Country" name="id" identifier="false"/>
+    <property resource="Sylius\Component\Addressing\Model\Country" name="code" identifier="true"/>
+</properties>
