--- src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/Country.xml
@@ -1,93 +1,0 @@
-<?xml version="1.0" ?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Sylius Sp. z o.o.
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<resources xmlns="https://api-platform.com/schema/metadata"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="https://api-platform.com/schema/metadata https://api-platform.com/schema/metadata/metadata-2.0.xsd"
->
-    <resource class="%sylius.model.country.class%" shortName="Country">
-        <attribute name="validation_groups">sylius</attribute>
-
-        <collectionOperations>
-            <collectionOperation name="admin_get">
-                <attribute name="method">GET</attribute>
-                <attribute name="path">/admin/countries</attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">admin:country:read</attribute>
-                </attribute>
-            </collectionOperation>
-
-            <collectionOperation name="shop_get">
-                <attribute name="method">GET</attribute>
-                <attribute name="path">/shop/countries</attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">shop:country:read</attribute>
-                </attribute>
-            </collectionOperation>
-
-            <collectionOperation name="admin_post">
-                <attribute name="method">POST</attribute>
-                <attribute name="path">/admin/countries</attribute>
-                <attribute name="denormalization_context">
-                    <attribute name="groups">admin:country:create</attribute>
-                </attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">admin:country:read</attribute>
-                </attribute>
-            </collectionOperation>
-        </collectionOperations>
-
-        <itemOperations>
-            <itemOperation name="admin_get">
-                <attribute name="method">GET</attribute>
-                <attribute name="path">/admin/countries/{code}</attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">admin:country:read</attribute>
-                </attribute>
-            </itemOperation>
-            <itemOperation name="shop_get">
-                <attribute name="method">GET</attribute>
-                <attribute name="path">/shop/countries/{code}</attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">shop:country:read</attribute>
-                </attribute>
-            </itemOperation>
-
-            <itemOperation name="admin_put">
-                <attribute name="method">PUT</attribute>
-                <attribute name="path">/admin/countries/{code}</attribute>
-                <attribute name="denormalization_context">
-                    <attribute name="groups">admin:country:update</attribute>
-                </attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">admin:country:read</attribute>
-                </attribute>
-            </itemOperation>
-        </itemOperations>
-
-        <subresourceOperations>
-            <subresourceOperation name="provinces_get_subresource">
-                <attribute name="method">GET</attribute>
-                <attribute name="path">/admin/countries/{code}/provinces</attribute>
-            </subresourceOperation>
-        </subresourceOperations>
-
-        <property name="id" identifier="false" writable="false" />
-        <property name="code" identifier="true" required="true" />
-        <property name="enabled" writable="true" />
-        <property name="name" writable="false" readable="true" />
-        <property name="provinces" readable="true" writable="true">
-            <subresource resourceClass="%sylius.model.province.class%" collection="true" />
-        </property>
-    </resource>
-</resources>
