--- src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/Province.xml
@@ -1,49 +1,0 @@
-<?xml version="1.0" ?>
-
-<!--
-
- This file is part of the Sylius package.
-
- (c) Sylius Sp. z o.o.
-
- For the full copyright and license information, please view the LICENSE
- file that was distributed with this source code.
-
--->
-
-<resources xmlns="https://api-platform.com/schema/metadata"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-           xsi:schemaLocation="https://api-platform.com/schema/metadata https://api-platform.com/schema/metadata/metadata-2.0.xsd"
->
-    <resource class="%sylius.model.province.class%" shortName="Province">
-        <attribute name="route_prefix">admin</attribute>
-
-        <attribute name="validation_groups">sylius</attribute>
-
-        <collectionOperations />
-
-        <itemOperations>
-            <itemOperation name="admin_get">
-                <attribute name="method">GET</attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">admin:province:read</attribute>
-                </attribute>
-            </itemOperation>
-
-            <itemOperation name="admin_put">
-                <attribute name="method">PUT</attribute>
-                <attribute name="denormalization_context">
-                    <attribute name="groups">admin:province:update</attribute>
-                </attribute>
-                <attribute name="normalization_context">
-                    <attribute name="groups">admin:province:read</attribute>
-                </attribute>
-            </itemOperation>
-        </itemOperations>
-
-        <property name="id" identifier="false" writable="false" />
-        <property name="code" identifier="true" required="true" />
-        <property name="name" required="true" />
-        <property name="abbreviation" writable="true" />
-    </resource>
-</resources>
