+++ src/Sylius/Bundle/ApiBundle/Resources/config/services.xml
@@ -15,9 +15,10 @@
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
 >
-<!--    <imports>-->
+    <imports>
 <!--        <import resource="services/**/*.xml" />-->
-<!--    </imports>-->
+        <import resource="services/state_processors.xml" />
+    </imports>
 
     <parameters>
         <parameter key="sylius.model.address.interface">Sylius\Component\Addressing\Model\AddressInterface</parameter>
